+++ tests/test_materialize.py
@@ -2,13 +2,14 @@ from __future__ import annotations
 
 import pandas as pd
 import pytest
-import sqlalchemy as sa
 
 from pydiverse.pipedag import Flow, Stage, Table, materialize
 from pydiverse.pipedag.context import RunContext, StageLockContext
 from pydiverse.pipedag.util.config import PipedagConfig
 
 from tests.util import select_as, tasks_library as m
+
+# parameterize all tests in this file with several instance_id configurations
 from tests.fixtures.instance import *